refactor: some changes for better reusability
