+++ src/appConstants/api.ts
@@ -1,0 +1,1 @@
+export const WEATHER_API_KEY = '0dd4c70fbcb17123e868e6d308f9906a';
